refactor: extract api implementation logic
