+++ core/data/src/main/java/jp/mydns/kokoichi0206/data/di/DataModule.kt
@@ -8,17 +8,11 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import jp.mydns.kokoichi0206.common.BuildConfigWrapper
 import jp.mydns.kokoichi0206.common.Constants
-import jp.mydns.kokoichi0206.common.interceptor.AddHeaderInterceptor
-import jp.mydns.kokoichi0206.common.interceptor.LoggingInterceptor
-import jp.mydns.kokoichi0206.common.interceptor.RetryInterceptor
 import jp.mydns.kokoichi0206.data.local.MembersDatabase
 import jp.mydns.kokoichi0206.data.local.QuizRecordDatabase
 import jp.mydns.kokoichi0206.data.remote.SakamichiApi
+import jp.mydns.kokoichi0206.data.remote.createSakamichiApi
 import jp.mydns.kokoichi0206.data.repository.*
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import java.util.concurrent.TimeUnit
 import javax.inject.Singleton
 
 /**
@@ -32,19 +26,7 @@ object DataModule {
     @Provides
     @Singleton
     fun provideSakamichiApi(): SakamichiApi {
-        val okHttpClient = OkHttpClient.Builder()
-            // サーバー側の設定か、なぜか指定が必要！
-            .connectTimeout(777, TimeUnit.MILLISECONDS)
-            .addInterceptor(AddHeaderInterceptor())
-            .addInterceptor(LoggingInterceptor())
-            .addInterceptor(RetryInterceptor())
-            .build()
-        return Retrofit.Builder()
-            .baseUrl(Constants.BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(MoshiConverterFactory.create().failOnUnknown())
-            .build()
-            .create(SakamichiApi::class.java)
+        return createSakamichiApi(Constants.BASE_URL)
     }
 
     @Provides
